requirements.txt
flask
pycryptodome
mysql-connector-python